refactor: Updated how marker is imported into cluster
